% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCrossingPercentage2.R
\name{calculateCrossingPercentage2}
\alias{calculateCrossingPercentage2}
\title{interpolates crossing time}
\usage{
calculateCrossingPercentage2(xi, yi, x1, y1, x2, y2, r)
}
\arguments{
\item{xCen}{x-coordinate of center of circle}

\item{yCen}{y-coordinate of center of circle}

\item{xIn}{x-coordinate of point inside circle}

\item{yIn}{y-coordinate of point inside circle}

\item{xOut}{x-coordinate of point outside circle}

\item{yOut}{y-coordinate of point outside circle}

\item{radius}{radius of circle}
}
\description{
used to verify alternate algorithm
}

