% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCrossingPercentage.R
\name{.calculateCrossingPercentageCmplx}
\alias{.calculateCrossingPercentageCmplx}
\title{Calculates percentage of trajectory segment within circle}
\usage{
.calculateCrossingPercentageCmplx(Cz, Az, Bz, R)
}
\arguments{
\item{Cz}{circle center}

\item{Az}{point 1}

\item{Bz}{point 2}

\item{R}{radius}
}
\description{
Calculates the percentage of a segment that lies within a circle for a point
A inside the circle and point B outside the circle for a circle with center C and radius R.
}
