% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBoundedEmpiricalCRW.R
\name{generateBoundedEmpiricalCRW}
\alias{generateBoundedEmpiricalCRW}
\title{Generates a bounded correlated random walk from an empirical set of steps and angles}
\usage{
generateBoundedEmpiricalCRW(n, startX, startY, steps, turnAngles, boundry)
}
\arguments{
\item{n}{number of steps to generate}

\item{startX}{starting x-coordiante for the walk}

\item{startY}{starting y-coordinate for the walk}

\item{steps}{vector of steps to draw from}

\item{turnAngles}{vector of turning angles to draw from}

\item{boundry}{a \link[sp]{SpatialPolygons} describing the boundary of the walk}
}
\value{
a data frame of the generated random walk with columns x and y describing the location 
and t for the time of each step
}
\description{
Generates a correlated random walk starting at (startX, startY) using pairs of steps and 
turning angles drawn from the provided empirical distributions. The random walk is bounded by a polygon by 
testing if each step will end outside the polygon and redrawing a new step if necessary.
}

