% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCRW.R
\name{generateCRW}
\alias{generateCRW}
\title{Generates a correlated random walk}
\usage{
generateCRW(n, weibull_shape = 2, weibull_scale = 1, cauchy_mu = 0,
  cauchy_rho = 0.8)
}
\arguments{
\item{n}{number of steps to generate}

\item{weibull_shape}{shape parameter of the Weibull distribution}

\item{weibull_scale}{scale parameter of the Weibull distribution}

\item{cauchy_mu}{mu parameter of the wrapped Cauchy distribution}

\item{cauchy_rho}{rho parameter of the wrapped Cauchy distribution}
}
\value{
a data frame of the generated random walk with columns x and y describing the location 
and t for the time of each step
}
\description{
Generates a correlated random walk starting at (0,0) using a Weibull distribution for step
length and a wrapped Cauchy distribution for turning angles.
}
\seealso{
\code{\link{rweibull}}, \code{\link[circular]{rwrappedcauchy}}
}

