% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateEmpiricalCRW.R
\name{generateEmpiricalCRW}
\alias{generateEmpiricalCRW}
\title{Generates a correlated random walk from an empirical set of steps and angles}
\usage{
generateEmpiricalCRW(n, startX, startY, steps, turnAngles)
}
\arguments{
\item{n}{number of steps to generate}

\item{startX}{starting x-coordiante for the walk}

\item{startY}{starting y-coordinate for the walk}

\item{steps}{vector of steps to draw from}

\item{turnAngles}{vector of turning angles to draw from}
}
\value{
a data frame of the generated random walk with columns x and y describing the location 
and t for the time of each step
}
\description{
Generates a correlated random walk starting at (startX, startY) using pairs of steps and 
turning angles drawn from the provided empirical distributions.
}

