% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurse.R
\docType{data}
\name{martin}
\alias{martin}
\title{Sample trajectory (martin).}
\format{
A data frame with 600 rows and 4 columns
}
\usage{
data(martin)
}
\description{
A dataset containing a sample trajectory with revisits.
}
\details{
\itemize{
  \item x. x-coordinate
  \item y. y-coordinate
  \item t. time
  \item id. identifier
}
}
\keyword{datasets}
