% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.recurse.R
\name{plot.recurse}
\alias{plot.recurse}
\title{Calculates recursion information from the trajectory}
\usage{
\method{plot}{recurse}(x, xyt, col, ...)
}
\arguments{
\item{x}{\code{recurse} object returned from call to \link{getRecursions}}

\item{xyt}{data.frame of x, y, and t representing the xy-coordinates and the time (same as call to \link{getRecursions})}

\item{col}{optional vector of colors as long as the maximum number of revisits to color code trajectory points}

\item{...}{additional arguments to \link{plot}}
}
\value{
the plot
}
\description{
Plots a trajectory color coded by number of revisits to each point.
}

