% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurse.R
\docType{package}
\name{recurse}
\alias{recurse}
\alias{recurse-package}
\title{Computes revisitation metrics for trajectory data}
\description{
Computes revisitation metrics for trajectory data, such as the number of revisitations for each location 
as well as the time spent for that visit and the time since the previous visit. Also includes functions to 
plot data and generate correlated random walks.
}
\details{
The function \code{\link{getRecursions}} computes the revisit metrics, which can be plotted with 
\code{\link{plot.recurse}}.

Functions are also available to generate correlated random walks from probability distributions or 
empirical distributions of steps and angles with \code{\link{generateCRW}} and 
\code{\link{generateEmpiricalCRW}}. The walk can also be bounded by a polygon (for example to keep it in
the correct habitat, land, water, etc.) with \code{\link{generateBoundedEmpiricalCRW}}.
}
\author{
Chloe Bracis <cbracis@uw.edu>
}

