% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurse.R
\docType{data}
\name{track}
\alias{track}
\title{Sample trajectory (track).}
\format{A data frame with 100 rows and 4 columns}
\usage{
data(track)
}
\description{
A dataset containing a sample trajectory with revisits.
}
\details{
\itemize{
  \item x. x-coordinate
  \item y. y-coordinate
  \item t. time
  \item id. identifier
}
}
\keyword{datasets}
