% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurse.R
\docType{package}
\name{recurse}
\alias{recurse}
\title{Computes revisitation metrics for trajectory data}
\description{
Computes revisitation metrics for trajectory data, such as the number of revisitations for each location 
as well as the time spent for that visit and the time since the previous visit. Also includes functions to 
plot data.
}
\details{
The function \code{\link{getRecursions}} computes the revisit metrics, which can be plotted with 
\code{\link{plot.recurse}}. Alternatively, \code{\link{getRecursionsAtLocations}} computes revisit metrics
for specified locations, rather than all locations in the movement trajectory.
}
\author{
Chloe Bracis <cbracis@uw.edu>
}
